add 'liveReloadBaseHost' option
In case your index.html is being served on a different domain
(i.e. your rails app that fetches it from redis)
the request for `ember-cli-live-reload.js` has to be triggered
towards the ember-cli host.

i.e.
if your rails app runs on `http://localhost:3000` the live reload
script should be requested at `http://localhost:4200/ember-cli-live-reload.js`.

The current config options do not seem to allow this setup
thanks to https://github.com/ember-cli/ember-cli/pull/4283 it is
possible to set the `host` for the liveReload server but this does
not eliminate the need to fetch `ember-cli-live-reload.js` from the
ember-cli server itself.

This change aims to solve this missing use case by appending the
`baseHost` to the script path